# Docs/OpenAPI.json
{
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Gateway service for Classroom microservices. Handles authentication, routing, and request aggregation.",
    "title": "Classroom Gateway API",
    "contact": {},
    "version": "1.0"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/lessons/create": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Создает новый урок в курсе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Lessons"],
        "summary": "Создание урока",
        "parameters": [
          {
            "description": "Данные для создания урока",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateLessonRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateLessonResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/lessons/lesson": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает детальную информацию об уроке",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Lessons"],
        "summary": "Получение урока",
        "parameters": [
          {
            "type": "string",
            "example": "\"94f9a22f-3a83-4591-a988-7aa3f0ec6eb0\"",
            "description": "ID урока",
            "name": "lesson_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetLessonResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Урок не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/lessons/lesson/update": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Обновляет информацию об уроке",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Lessons"],
        "summary": "Обновление урока",
        "parameters": [
          {
            "description": "Данные для обновления",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateLessonRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateLessonResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Урок не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/lessons/lesson/delete": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Удаляет урок по идентификатору",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Lessons"],
        "summary": "Удаление урока",
        "parameters": [
          {
            "description": "Идентификатор урока",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteLessonRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteLessonResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Урок не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/lessons/lessons": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список уроков с возможностью фильтрации по курсу",
        "produces": ["application/json"],
        "tags": ["Lessons"],
        "summary": "Получение списка уроков",
        "parameters": [
          {
            "type": "string",
            "example": "\"6994aefe-6815-476b-bdc0-2ae5c4d0c18e\"",
            "description": "ID курса",
            "name": "course_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetLessonsResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateLessonRequest": {
      "description": "Параметры для создания нового занятия в курсе",
      "type": "object",
      "properties": {
        "course_id": {
          "description": "ID курса",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        },
        "title": {
          "description": "Название занятия",
          "type": "string",
          "x-order": "1",
          "example": "Основы алгоритмов"
        },
        "content": {
          "description": "Содержание занятия",
          "type": "string",
          "x-order": "2",
          "example": "Подробное описание занятия..."
        }
      }
    },
    "CreateLessonResponse": {
      "description": "Возвращает ID созданного занятия",
      "type": "object",
      "properties": {
        "lesson_id": {
          "description": "ID созданного занятия",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        }
      }
    },
...
    "DeleteLessonRequest": {
      "description": "Требует ID курса и ID занятия для удаления",
      "type": "object",
      "properties": {
        "lesson_id": {
          "description": "ID занятия",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        }
      }
    },
    "DeleteLessonResponse": {
      "description": "Пустой ответ при успешном удалении",
      "type": "object"
    },
...
    "ErrorResponse": {
      "description": "Cтандартизированный формат всех ошибок в API.",
      "type": "object",
      "properties": {
        "code": {
          "description": "Код ошибки (HTTP статус код)",
          "type": "integer",
          "example": 400
        },
        "message": {
          "description": "Сообщение об ошибке",
          "type": "string",
          "example": "Bad request"
        }
      }
    },
...
    "GetLessonResponse": {
      "description": "Возвращает полные данные занятия",
      "type": "object",
      "properties": {
        "lesson": {
          "description": "Объект занятия",
          "allOf": [
            {
              "$ref": "#/definitions/Lesson"
            }
          ],
          "x-order": "0"
        }
      }
    },
    "GetLessonsResponse": {
      "description": "Содержит массив занятий в курсе",
      "type": "object",
      "properties": {
        "lessons": {
          "description": "Массив занятий",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Lesson"
          },
          "x-order": "0"
        }
      }
    },
...
    "Lesson": {
      "description": "Полная информация о занятии в курсе",
      "type": "object",
      "properties": {
        "lesson_id": {
          "description": "Уникальный идентификатор занятия",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        },
        "course_id": {
          "description": "Идентификатор курса",
          "type": "string",
          "x-order": "1",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        },
        "title": {
          "description": "Название занятия",
          "type": "string",
          "x-order": "2",
          "example": "Введение в программирование"
        },
        "description": {
          "description": "Содержание занятия",
          "type": "string",
          "x-order": "3",
          "example": "Базовые понятия и термины"
        },
        "created_at": {
          "description": "Дата создания",
          "type": "string",
          "x-order": "4",
          "example": "2023-01-15T10:00:00Z"
        }
      }
    },
...
    "UpdateLessonRequest": {
      "description": "Позволяет частично обновить данные занятия",
      "type": "object",
      "properties": {
        "lesson_id": {
          "description": "ID занятия",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        },
        "title": {
          "description": "Новое название (опционально)",
          "type": "string",
          "x-order": "1",
          "example": "Обновленное название"
        },
        "description": {
          "description": "Новое содержание (опционально)",
          "type": "string",
          "x-order": "2",
          "example": "Обновленное содержание"
        }
      }
    },
    "UpdateLessonResponse": {
      "description": "Пустой ответ при успешном обновлении",
      "type": "object",
      "properties": {
        "lesson": {
          "allOf": [
            {
              "$ref": "#/definitions/Lesson"
            }
          ],
          "x-order": "0"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Enter JWT token in format: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}